backend/requirements.txt
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
web3>=6.15.0
eth-account>=0.11.0
